examples/crewai/requirements_enhanced.txt
maif[integrations]
crewai
chromadb
sentence-transformers
google-generativeai
python-dotenv
tqdm
requests


